+++ apps/help-center/package.json
@@ -25,13 +25,19 @@
 		"show-stats": "NODE_ENV=production EMIT_STATS=true yarn build"
 	},
 	"dependencies": {
+		"@automattic/calypso-polyfills": "workspace:^",
+		"@automattic/data-stores": "workspace:^",
 		"@automattic/happychat-connection": "workspace:^",
 		"@automattic/help-center": "workspace:^",
 		"@automattic/site-picker": "workspace:^",
+		"@wordpress/components": "^19.9.0",
+		"@wordpress/compose": "^5.5.0",
+		"@wordpress/data": "^6.7.0",
+		"@wordpress/element": "^4.5.0",
 		"@wordpress/interface": "^4.6.0",
 		"@wordpress/plugins": "^4.5.0",
-		"@automattic/calypso-polyfills": "workspace:^",
 		"calypso": "workspace:^",
+		"classnames": "^2.3.1",
 		"react": "^17.0.2",
 		"react-dom": "^17.0.2",
 		"react-query": "^3.32.1",
