+++ apps/help-center/src/index.js
@@ -1,23 +1,66 @@
-import { PluginSidebar, PluginSidebarMoreMenuItem } from '@wordpress/edit-post';
+/* eslint-disable wpcalypso/jsx-classname-namespace */
+import './config';
+import { useHasSeenWhatsNewModalQuery } from '@automattic/data-stores';
+import HelpCenter, { HelpIcon } from '@automattic/help-center';
+import { Button } from '@wordpress/components';
+import { useMediaQuery } from '@wordpress/compose';
+import { useDispatch, useSelect } from '@wordpress/data';
 import { PinnedItems } from '@wordpress/interface';
 import { registerPlugin } from '@wordpress/plugins';
+import cx from 'classnames';
+import { useEffect, useState } from 'react';
+import { QueryClientProvider } from 'react-query';
+import { whatsNewQueryClient } from '../../editing-toolkit/editing-toolkit-plugin/common/what-new-query-client';
+import Contents from './contents';
+import './help-center.scss';
+
+function HelpCenterContent() {
+	const isDesktop = useMediaQuery( '(min-width: 480px)' );
+	const show = useSelect( ( select ) => select( 'automattic/help-center' ).isHelpCenterShown() );
+	const { setShowHelpCenter } = useDispatch( 'automattic/help-center' );
+	const [ showHelpIconDot, setShowHelpIconDot ] = useState( false );
+	const { data, isLoading } = useHasSeenWhatsNewModalQuery( window._currentSiteId );
+	useEffect( () => {
+		if ( ! isLoading && data ) {
+			setShowHelpIconDot( ! data.has_seen_whats_new_modal );
+		}
+	}, [ data, isLoading ] );
+
+	const content = (
+		<span className="help-center-entry-point">
+			<Button
+				className={ cx( 'entry-point-button', { 'is-active': show } ) }
+				onClick={ () => setShowHelpCenter( ! show ) }
+				icon={ <HelpIcon newItems={ showHelpIconDot } active={ show } /> }
+			></Button>
+		</span>
+	);
+
+	return (
+		<>
+			{ isDesktop && (
+				<>
+					<PinnedItems scope="core/edit-post">{ content }</PinnedItems>
+					<PinnedItems scope="core/edit-site">{ content }</PinnedItems>
+					<PinnedItems scope="core/edit-widgets">{ content }</PinnedItems>
+				</>
+			) }
+			{ show && (
+				<HelpCenter
+					content={ <Contents handleClose={ () => setShowHelpCenter( false ) } /> }
+					handleClose={ () => setShowHelpCenter( false ) }
+				/>
+			) }
+		</>
+	);
+}
 
 registerPlugin( 'etk-help-center', {
 	render: () => {
-		debugger;
 		return (
-			<PluginSidebarMoreMenuItem
-				icon={
-					<svg width="20" viewBox="0 0 20 20">
-						<circle r="10" x="10" y="10" />
-					</svg>
-				}
-				target="global-styles"
-			>
-				sss
-			</PluginSidebarMoreMenuItem>
+			<QueryClientProvider client={ whatsNewQueryClient }>
+				<HelpCenterContent />,
+			</QueryClientProvider>
 		);
 	},
 } );
-
-console.log( 'omar4' );
