+++ apps/help-center/src/store.js
@@ -1,0 +1,5 @@
+import { HelpCenter, Site, User } from '@automattic/data-stores';
+
+export const HELP_CENTER_STORE = HelpCenter.register();
+export const SITE_STORE = Site.register();
+export const USER_STORE = User.register();
